app/requirements.txt
argon2==0.1.10
argon2-cffi==20.1.0
dnspython==1.16.0
fastapi==0.61.0
fasteners==0.14.1
gunicorn==20.0.4
html5lib==1.0.1
httplib2==0.14.0
httptools==0.1.1
pandas==1.1.1
psycopg2-binary==2.8.5
pydantic==1.6.1
PyJWT==1.7.1
pymongo==3.11.0
requests==2.22.0
starlette==0.13.6
uvicorn==0.11.8
virtualenv==20.0.17
wrapt==1.12.1
python-dotenv 0.14.0 
